Use autoload to delay loading features until we need them.
This yields about a 10% improvement in the time it
takes to initially load rspec-mocks.
