+++ lib/rspec/mocks/matchers/receive.rb
@@ -1,3 +1,5 @@
+require 'rspec/mocks/matchers/expectation_customization'
+
 module RSpec
   module Mocks
     module Matchers
@@ -99,21 +101,6 @@ module RSpec
           last.block ||= block
           nil
         end
-      end
-    end
-
-    # @private
-    class ExpectationCustomization
-      attr_accessor :block
-
-      def initialize(method_name, args, block)
-        @method_name = method_name
-        @args        = args
-        @block       = block
-      end
-
-      def playback_onto(expectation)
-        expectation.__send__(@method_name, *@args, &@block)
       end
     end
   end
